move validation policy code to cadc-storage-adapter add StorageAdapter.getValidationPolicy() remove redundant test code from tantar
