+++ cadc-storage-adapter/build.gradle
@@ -14,7 +14,7 @@ sourceCompatibility = 1.8
 
 group = 'org.opencadc'
 
-version = '0.10.0'
+version = '0.10.1'
 
 description = 'OpenCADC Storage Inventory StorageAdapter API library'
 def git_url = 'https://github.com/opencadc/storage-inventory'
