+++ cadc-storage-adapter/src/main/java/org/opencadc/inventory/storage/policy/RecoverFromStorage.java
@@ -1,9 +1,10 @@
+
 /*
  ************************************************************************
  *******************  CANADIAN ASTRONOMY DATA CENTRE  *******************
  **************  CENTRE CANADIEN DE DONNÉES ASTRONOMIQUES  **************
  *
- *  (c) 2020.                            (c) 2020.
+ *  (c) 2022.                            (c) 2022.
  *  Government of Canada                 Gouvernement du Canada
  *  National Research Council            Conseil national de recherches
  *  Ottawa, Canada, K1A 0R6              Ottawa, Canada, K1A 0R6
@@ -66,80 +67,70 @@
  ************************************************************************
  */
 
-package org.opencadc.tantar;
+package org.opencadc.inventory.storage.policy;
 
-import java.net.URI;
-import java.util.EventListener;
-
+import org.apache.log4j.Logger;
 import org.opencadc.inventory.Artifact;
 import org.opencadc.inventory.storage.StorageMetadata;
 
+/**
+ * Policy to ensure that a recovery from Storage (in the event of a disaster or a new site is brought online) will
+ * dictate what goes into the Inventory Database. This policy differs from StorageIsAlwaysRight because... ???
+ */
+public class RecoverFromStorage extends StorageValidationPolicy {
+    private static final Logger log = Logger.getLogger(RecoverFromStorage.class);
 
-public interface ValidateEventListener extends EventListener {
-    
-    /**
-     * Get existing artifact by uri. This is to support possible changes in StorageLocation
-     * for recovery of existing artifact-storage connection.
-     * 
-     * @param uri
-     * @return the artifact or null
-     */
-    Artifact getArtifact(URI uri);
+    public RecoverFromStorage(ValidateEventListener validateEventListener) {
+        super(validateEventListener);
+        throw new UnsupportedOperationException("** incomplete implementation **");
+    }
 
     /**
-     * Create a new Artifact using metadata from the given StorageMetadata.
+     * Use the logic of this Policy to correct a conflict caused by the two given items.  Only the Artifact can be
+     * null; it will be assumed that the given StorageMetadata is never null.
      *
-     * @param storageMetadata       The StorageMetadata to pull metadata from.
-     * @throws Exception    Any unexpected error.
+     * @param artifact        The Artifact to use in deciding.  Can be null.
+     * @param storageMetadata The StorageMetadata to use in deciding.  Never null.
+     * @throws Exception For any unknown error that should be passed up.
      */
-    void createArtifact(final StorageMetadata storageMetadata) throws Exception;
-
-    /**
-     * Delete the given StorageMetadata.
-     *
-     * @param storageMetadata   The StorageMetadata to delete
-     * @throws Exception    Any unexpected error.
-     */
-    void delete(final StorageMetadata storageMetadata) throws Exception;
-
-    /**
-     * Delete the given Artifact.  Implementors should also create a DeletedArtifactEvent as necessary.
-     *
-     * @param artifact      The Artifact to remove.
-     * @throws Exception    Any unexpected error.
-     */
-    void delete(final Artifact artifact) throws Exception;
-
-    /**
-     * This will force the file-sync application to assume it's a new insert and force a re-download of the file.  The
-     * default logic will most likely be to remove its StorageLocation instance.
-     *
-     * @param artifact The base artifact.  This MUST have a Storage Location.
-     * @throws Exception Anything IO/Thread related.
-     */
-    void clearStorageLocation(final Artifact artifact) throws Exception;
-
-    /**
-     * Replace the given Artifact with a new one created from the given StorageMetadata instance.
-     *
-     * @param artifact          The Artifact to replace.
-     * @param storageMetadata   The StorageMetadata from which to create a new Artifact.
-     * @throws Exception    Any unexpected error.
-     */
-    void replaceArtifact(final Artifact artifact, final StorageMetadata storageMetadata) throws Exception;
-
-    /**
-     * Update the values of the given Artifact with those from the given StorageMetadata.  This differs from a replace
-     * as it will not delete the original Artifact first, but rather update the values and issue a PUT.
-     *
-     * @param artifact          The Artifact to update.
-     * @param storageMetadata   The StorageMetadata from which to update the Artifact's fields.
-     * @throws Exception    Any unexpected error.
-     */
-    void updateArtifact(final Artifact artifact, final StorageMetadata storageMetadata) throws Exception;
-    
-    /**
-     * Delay validation but increment count.
-     */
-    void delayAction();
+    @Override
+    public void validate(final Artifact artifact, final StorageMetadata storageMetadata) throws Exception {
+        if (artifact == null && storageMetadata == null) {
+            throw new RuntimeException("BUG: both args to resolve are null");
+        }
+        
+        StringBuilder sb = new StringBuilder();
+        sb.append(this.getClass().getSimpleName());
+        if (storageMetadata == null || !storageMetadata.isValid()) {
+            sb.append(".noAction");
+            if (artifact != null) {
+                sb.append(" Artifact.id=").append(artifact.getID());
+                sb.append(" Artifact.uri=").append(artifact.getURI());
+                sb.append(" reason=no-matching-storageLocation");
+            } else {
+                sb.append(" reason=invalid-storageLocation");
+            }
+            log.info(sb.toString());
+            return;
+        }
+        
+        if (artifact == null) {
+            sb.append(".createArtifact");
+            sb.append(" Artifact.uri=").append(storageMetadata.getArtifactURI());
+            sb.append(" loc=").append(storageMetadata.getStorageLocation());
+            //log.info(String.format("Adding Artifact %s as per policy.", storageMetadata.getStorageLocation()));
+            log.info(sb.toString());
+            validateEventListener.createArtifact(storageMetadata);
+        } else  {
+            sb.append(".updateArtifact");
+            sb.append(" Artifact.id=").append(artifact.getID());
+            sb.append(" Artifact.uri=").append(artifact.getURI());
+            sb.append(" loc=").append(storageMetadata.getStorageLocation());
+            // This scenario is for an incomplete previous run.  Treat the Artifact as corrupt and set it back to the
+            // StorageMetadata's values.
+            //log.info(String.format("Updating Artifact %s as per policy.", storageMetadata.getStorageLocation()));
+            log.info(sb.toString());
+            validateEventListener.updateArtifact(artifact, storageMetadata);
+        }
+    }
 }
