+++ cadc-storage-adapter/src/test/java/org/opencadc/inventory/storage/policy/StorageIsAlwaysRightTest.java
@@ -4,7 +4,7 @@
  *******************  CANADIAN ASTRONOMY DATA CENTRE  *******************
  **************  CENTRE CANADIEN DE DONNÉES ASTRONOMIQUES  **************
  *
- *  (c) 2020.                            (c) 2020.
+ *  (c) 2022.                            (c) 2022.
  *  Government of Canada                 Gouvernement du Canada
  *  National Research Council            Conseil national de recherches
  *  Ottawa, Canada, K1A 0R6              Ottawa, Canada, K1A 0R6
@@ -67,7 +67,7 @@
  ************************************************************************
  */
 
-package org.opencadc.tantar.policy;
+package org.opencadc.inventory.storage.policy;
 
 import java.io.ByteArrayOutputStream;
 import java.net.URI;
@@ -79,15 +79,12 @@ import org.junit.Test;
 import org.opencadc.inventory.Artifact;
 import org.opencadc.inventory.StorageLocation;
 import org.opencadc.inventory.storage.StorageMetadata;
-import org.opencadc.tantar.Reporter;
-import org.opencadc.tantar.TestEventListener;
 
 public class StorageIsAlwaysRightTest extends AbstractResolutionPolicyTest<StorageIsAlwaysRight> {
 
     @Test
     public void resolveArtifactAndStorageMetadata() throws Exception {
         final ByteArrayOutputStream output = new ByteArrayOutputStream();
-        final Reporter reporter = new Reporter(getTestLogger(output));
         final Artifact artifact = new Artifact(URI.create("cadc:bucket/file.fits"),
                                                URI.create("md5:" + random16Bytes()), new Date(), 88L);
         artifact.storageLocation = new StorageLocation(URI.create("s3:101010"));
@@ -96,8 +93,8 @@ public class StorageIsAlwaysRightTest extends AbstractResolutionPolicyTest<Stora
                                                         URI.create("md5:" + random16Bytes()), 1001L, new Date());
         final TestEventListener testEventListener = new TestEventListener();
 
-        testSubject = new StorageIsAlwaysRight(testEventListener, reporter);
-        testSubject.resolve(artifact, storageMetadata);
+        testSubject = new StorageIsAlwaysRight(testEventListener);
+        testSubject.validate(artifact, storageMetadata);
 
         final List<String> outputLines = Arrays.asList(new String(output.toByteArray()).split("\n"));
         System.out.println(String.format("Message lines are \n\n%s\n\n", outputLines));
@@ -114,15 +111,14 @@ public class StorageIsAlwaysRightTest extends AbstractResolutionPolicyTest<Stora
     @Test
     public void resolveNullAndInvalidStorageMetadata() throws Exception {
         final ByteArrayOutputStream output = new ByteArrayOutputStream();
-        final Reporter reporter = new Reporter(getTestLogger(output));
 
         // StorageMetadata nas no other metadata than the StorageLocation.
         final StorageMetadata storageMetadata = new StorageMetadata(new StorageLocation(URI.create("s3:989877")));
 
         final TestEventListener testEventListener = new TestEventListener();
 
-        testSubject = new StorageIsAlwaysRight(testEventListener, reporter);
-        testSubject.resolve(null, storageMetadata);
+        testSubject = new StorageIsAlwaysRight(testEventListener);
+        testSubject.validate(null, storageMetadata);
 
         final List<String> outputLines = Arrays.asList(new String(output.toByteArray()).split("\n"));
         System.out.println(String.format("Message lines are \n\n%s\n\n", outputLines));
@@ -142,15 +138,15 @@ public class StorageIsAlwaysRightTest extends AbstractResolutionPolicyTest<Stora
     @Test
     public void resolveArtifactAndNull() throws Exception {
         final ByteArrayOutputStream output = new ByteArrayOutputStream();
-        final Reporter reporter = new Reporter(getTestLogger(output));
+
         final Artifact artifact = new Artifact(URI.create("cadc:bucket/file.fits"),
                                                URI.create("md5:" + random16Bytes()), new Date(), 88L);
         final TestEventListener testEventListener = new TestEventListener();
 
         artifact.storageLocation = new StorageLocation(URI.create("s3:101010"));
 
-        testSubject = new StorageIsAlwaysRight(testEventListener, reporter);
-        testSubject.resolve(artifact, null);
+        testSubject = new StorageIsAlwaysRight(testEventListener);
+        testSubject.validate(artifact, null);
 
         final List<String> outputLines = Arrays.asList(new String(output.toByteArray()).split("\n"));
         System.out.println(String.format("Message lines are \n\n%s\n\n", outputLines));
@@ -166,7 +162,6 @@ public class StorageIsAlwaysRightTest extends AbstractResolutionPolicyTest<Stora
     @Test
     public void resolveArtifactAndInvalidStorageMetadata() throws Exception {
         final ByteArrayOutputStream output = new ByteArrayOutputStream();
-        final Reporter reporter = new Reporter(getTestLogger(output));
 
         // StorageMetadata nas no other metadata than the StorageLocation.
         final StorageMetadata storageMetadata = new StorageMetadata(new StorageLocation(URI.create("s3:989877")));
@@ -179,8 +174,8 @@ public class StorageIsAlwaysRightTest extends AbstractResolutionPolicyTest<Stora
 
         final TestEventListener testEventListener = new TestEventListener();
 
-        testSubject = new StorageIsAlwaysRight(testEventListener, reporter);
-        testSubject.resolve(artifact, storageMetadata);
+        testSubject = new StorageIsAlwaysRight(testEventListener);
+        testSubject.validate(artifact, storageMetadata);
 
         final List<String> outputLines = Arrays.asList(new String(output.toByteArray()).split("\n"));
         System.out.println(String.format("Message lines are \n\n%s\n\n", outputLines));
@@ -200,13 +195,13 @@ public class StorageIsAlwaysRightTest extends AbstractResolutionPolicyTest<Stora
     @Test
     public void resolveNullAndStorageMetadata() throws Exception {
         final ByteArrayOutputStream output = new ByteArrayOutputStream();
-        final Reporter reporter = new Reporter(getTestLogger(output));
+
         final StorageMetadata storageMetadata = new StorageMetadata(new StorageLocation(URI.create("s3:101011")), URI.create("test:101011"),
                                                         URI.create("md5:" + random16Bytes()), 1001L, new Date());
         final TestEventListener testEventListener = new TestEventListener();
 
-        testSubject = new StorageIsAlwaysRight(testEventListener, reporter);
-        testSubject.resolve(null, storageMetadata);
+        testSubject = new StorageIsAlwaysRight(testEventListener);
+        testSubject.validate(null, storageMetadata);
 
         final List<String> outputLines = Arrays.asList(new String(output.toByteArray()).split("\n"));
         System.out.println(String.format("Message lines are \n\n%s\n\n", outputLines));
